Move a sample file over to the new libOPHD project
