+++ OPHD/ophd.vcxproj
@@ -338,7 +338,6 @@
     <ClInclude Include="ProductCatalogue.h" />
     <ClInclude Include="ProductionCost.h" />
     <ClInclude Include="ProductPool.h" />
-    <ClInclude Include="RandomNumberGenerator.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="RobotPool.h" />
     <ClInclude Include="ShellOpenPath.h" />
