+++ libOPHD/RandomNumberGenerator.h
